Test: Run state coverage on Linux
- move state and storage units into TUIkitViewTests
- replace the Linux-wide skip with serialized portable tests
- keep render hydration behavior in the umbrella integration suite
